\name{updateDescription}
\alias{updateDescription}
\title{ Set the description on a TIFF file. }
\description{
  Sets the TIFF description field (tag 270) for a TIFF file.
}
\usage{
updateDescription(fn, description)
}
\arguments{
  \item{fn}{ The name of the tiff file. }
  \item{description}{ The text to set the description to. }
}
\value{
None.
}
\author{ Eric Kort <eric.kort@vai.org> }
\keyword{ utilities }
\keyword{ IO }
\keyword{ file }
